voice_assistant/requirements.txt
gspread==6.1.2
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.114.0
SpeechRecognition==3.10.3
python-dotenv==1.0.1
duckduckgo-search==6.1.12
openai==1.35.10

# Optional: install pyaudio for live microphone capture (system package needed on macOS/Linux)
